Refactor DB schema and accessor functions

+++ bobby/cass.py
@@ -22,11 +22,11 @@ class ResultNotFoundError(Exception):
 
 def get_groups_by_tenant_id(tenant_id):
     """Get all groups owned by a provided tenant."""
-    query = 'SELECT * FROM groups WHERE "tenantId"=:tenantId ALLOW FILTERING;'
+    query = 'SELECT * FROM groups WHERE "tenantId"=:tenantId;'
     return _client.execute(query, {'tenantId': tenant_id}, ConsistencyLevel.ONE)
 
 
-def get_group_by_id(group_id):
+def get_group_by_id(tenant_id, group_id):
     """Get a group by its id."""
     query = 'SELECT * FROM groups WHERE "groupId"=:groupId;'
     d = _client.execute(query, {'groupId': group_id}, ConsistencyLevel.ONE)
@@ -56,31 +56,33 @@ def create_group(group_id, tenant_id, notification, notification_plan):
     d = _client.execute(query, data, ConsistencyLevel.ONE)
 
     def retrieve_new_group(_):
-        return get_group_by_id(group_id)
+        return get_group_by_id(tenant_id, group_id)
     return d.addCallback(retrieve_new_group)
 
 
-def delete_group(group_id):
+def delete_group(tenant_id, group_id):
     """Delete a group."""
-    query = 'DELETE FROM groups WHERE "groupId"=:groupId;'
+    query = 'DELETE FROM groups WHERE "groupId"=:groupId AND "tenantId"=:tenantId;'
     return _client.execute(query,
-                           {'groupId': group_id},
+                           {'groupId': group_id,
+                           'tenantId': tenant_id},
                            ConsistencyLevel.ONE)
 
 
-def get_servers_by_group_id(group_id):
+def get_servers_by_group_id(tenant_id, group_id):
     """Get all servers with a specified groupId."""
-    query = 'SELECT * FROM servers WHERE "groupId"=:groupId ALLOW FILTERING;'
+    query = 'SELECT * FROM servers WHERE "groupId"=:groupId AND "tenantId"=:tenantId;'
 
-    return _client.execute(query, {'groupId': group_id}, ConsistencyLevel.ONE)
+    return _client.execute(query, {'groupId': group_id, 'tenantId': tenant_id},
+                           ConsistencyLevel.ONE)
 
 
-def get_server_by_server_id(server_id):
+def get_server_by_server_id(tenant_id, server_id):
     """Get a server by its serverId."""
 
-    query = 'SELECT * FROM servers WHERE "serverId"=:serverId;'
+    query = 'SELECT * FROM servers WHERE "serverId"=:serverId AND "tenantId"=:tenantId;'
 
-    d = _client.execute(query, {'serverId': server_id},
+    d = _client.execute(query, {'serverId': server_id, 'tenantId': tenant_id},
                         ConsistencyLevel.ONE)
 
     def return_server(result):
@@ -92,72 +94,45 @@ def get_server_by_server_id(server_id):
     return d.addCallback(return_server)
 
 
-def create_server(server_id, entity_id, group_id, server_policies):
+def create_server(tenant_id, server_id, entity_id, group_id):
     """Create and return a new server dict."""
     query = ' '.join([
-        'INSERT INTO servers ("serverId", "entityId", "groupId")',
-        'VALUES (:serverId, :entityId, :groupId);'])
+        'INSERT INTO servers ("tenantId", "serverId", "entityId", "groupId")',
+        'VALUES (:tenantId, :serverId, :entityId, :groupId);'])
 
     d = _client.execute(query,
-                        {'serverId': server_id, 'entityId': entity_id, 'groupId': group_id},
+                        {'serverId': server_id, 'entityId': entity_id, 'groupId': group_id,
+                         'tenantId': tenant_id},
                         ConsistencyLevel.ONE)
 
-    def add_server_policies(_):
-        query = ' '.join([
-            'INSERT INTO serverpolicies ("serverId", "policyId",',
-            '"alarmId", "checkId", "state") VALUES (:serverId, :policyId,',
-            ':alarmId, :checkId, :state);'])
-        deferreds = []
-        for server_policy in server_policies:
-            d = _client.execute(query, {'serverId': server_id,
-                                        'policyId': server_policy['policyId'],
-                                        'alarmId': server_policy['alarmId'],
-                                        'checkId': server_policy['checkId'],
-                                        'state': 'OK'},
-                                ConsistencyLevel.ONE)
-            deferreds.append(d)
-        return defer.DeferredList(deferreds)
-    d.addCallback(add_server_policies)
-
     def retrieve_server(_):
-        return get_server_by_server_id(server_id)
+        return get_server_by_server_id(tenant_id, server_id)
     return d.addCallback(retrieve_server)
 
 
-def delete_server(server_id):
+def delete_server(tenant_id, server_id):
     """Delete a server and cascade to deleting related serverpolicies."""
     # TODO: also delete the entity is MaaS
-    query = 'DELETE FROM servers WHERE "serverId"=:serverId;'
+    query = 'DELETE FROM servers WHERE "serverId"=:serverId AND "tenantId"=:tenantId;'
     d = _client.execute(query,
-                        {'serverId': server_id},
+                        {'serverId': server_id, 'tenantId': tenant_id},
                         ConsistencyLevel.ONE)
 
-    def remove_server_policies(_):
-        query = 'DELETE FROM serverpolicies WHERE "serverId"=:serverId;'
-        return _client.execute(query, {'serverId': server_id}, ConsistencyLevel.ONE)
-    return d.addCallback(remove_server_policies)
-
-
-def get_serverpolicies_for_server(server_id):
-    """Get all server policies for a provided server_id."""
-    query = 'SELECT * FROM serverpolicies WHERE "serverId"=:serverId;'
-    return _client.execute(query,
-                           {'serverId': server_id},
-                           ConsistencyLevel.ONE)
+    return d
 
 
 def get_policies_by_group_id(group_id):
     """Get all policies owned by a provided groupId."""
-    query = 'SELECT * FROM policies WHERE "groupId"=:groupId ALLOW FILTERING;'
+    query = 'SELECT * FROM policies WHERE "groupId"=:groupId;'
     return _client.execute(query, {'groupId': group_id},
                            ConsistencyLevel.ONE)
 
 
-def get_policy_by_policy_id(policy_id):
+def get_policy_by_policy_id(group_id, policy_id):
     """Get a single policy by its policyId."""
-    query = 'SELECT * FROM policies WHERE "policyId"=:policyId ALLOW FILTERING;'
+    query = 'SELECT * FROM policies WHERE "policyId"=:policyId AND "groupId"=:groupId;'
 
-    d = _client.execute(query, {'policyId': policy_id}, ConsistencyLevel.ONE)
+    d = _client.execute(query, {'policyId': policy_id, 'groupId': group_id}, ConsistencyLevel.ONE)
 
     def return_policy(result):
         if len(result) < 1:
@@ -183,7 +158,7 @@ def create_policy(policy_id, group_id, alarm_template, check_template):
                         ConsistencyLevel.ONE)
 
     def retrieve_policy(_):
-        return get_policy_by_policy_id(policy_id)
+        return get_policy_by_policy_id(group_id, policy_id)
     return d.addCallback(retrieve_policy)
 
 
@@ -195,12 +170,42 @@ def delete_policy(policy_id):
                         {'policyId': policy_id},
                         ConsistencyLevel.ONE)
 
-    # TODO: re-enable this. Cassandra's composite primary keys are a *real*
-    # head scratcher.
-    #def remove_server_policies(_):
-    #    query = 'DELETE FROM serverpolicies WHERE "policyId"=:policyId;'
-    #    return _client.execute(query, {'policyId': policy_id}, ConsistencyLevel.ONE)
     return d
+
+
+def register_policy_on_server(policy_id, server_id, alarm_id, check_id):
+    """Create a serverpolicy."""
+    query = ' '.join((
+        'INSERT INTO serverpolicies ("serverId", "policyId", "alarmId", "checkId", state)',
+        'VALUES (:serverId, :policyId, :alarmId, :checkId, false);'
+    ))
+
+    d = _client.execute(query,
+                        {'policyId': policy_id,
+                         'serverId': server_id,
+                         'alarmId': alarm_id,
+                         'checkId': check_id},
+                        ConsistencyLevel.ONE)
+    return d
+
+
+def deregister_policy_on_server(policy_id, server_id):
+    """Delete a serverpolicy record."""
+    query = 'DELETE FROM serverpolicies WHERE "policyId"=:policyId AND "serverId"=:serverId;'
+
+    d = _client.execute(query,
+                        {'policyId': policy_id,
+                         'serverId': server_id},
+                        ConsistencyLevel.ONE)
+
+    return d
+
+
+def get_policy_state(policy_id):
+    """ Get the state of the policy checks on each server. """
+    query = 'SELECT * FROM serverpolicies WHERE "policyId"=:policyId;'
+    return _client.execute(query, {'policyId': policy_id},
+                           ConsistencyLevel.ONE)
 
 
 _client = None
